Fix lint issues and move TransactionWithOrganization interface
- Move TransactionWithOrganization to get-transactions-usecase.ts
- Fix eslint-disable comment placement
- Update imports to use absolute paths
