--- admin/src/shared/models/transaction-with-organization.ts
@@ -1,5 +1,0 @@
-import type { Transaction } from "@/shared/models/transaction";
-
-export interface TransactionWithOrganization extends Transaction {
-  political_organization_name?: string;
-}
